refractor(course-detail): folder structure
